refactor: moved product utils to utils directory
